exception handled.

+++ HibernateDemo/src/main/java/org/nahidsohel/DeleteApp.java
@@ -7,25 +7,26 @@ import org.nahidsohel.entity.Song;
 import org.nahidsohel.utils.HibernateUtils;
 
 public class DeleteApp {
-
-    public static void main( String[] args ){
+	
+	public static void main(String[] args) {
 
 //        Configuration configuration = new Configuration();
 //        configuration.configure("hibernate.cfg.xml");
 //        configuration.addAnnotatedClass(Song.class);
 
 //        Create SessionFactory
-        SessionFactory sessionFactory = HibernateUtils.getSessionFactory();
-
-        Session session = sessionFactory.openSession();
-
-        Song song = session.get(Song.class,13);
-        System.out.println(song);
-
-        session.beginTransaction();
-        session.delete(song);
-        session.getTransaction().commit();
-
-        session.close();
-    }
+		SessionFactory sessionFactory = HibernateUtils.getSessionFactory();
+		if (sessionFactory != null) {
+			Session session = sessionFactory.openSession();
+			
+			Song song = session.get(Song.class, 13);
+			System.out.println(song);
+			
+			session.beginTransaction();
+			session.delete(song);
+			session.getTransaction().commit();
+			
+			session.close();
+		}
+	}
 }
